add router protected

+++ controllers/videoController.js
@@ -41,8 +41,11 @@ export const postUploadVideo = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   })
+  req.user.videos.push(newVideo._id)
+  req.user.save()
   res.redirect(routes.videoDetail(newVideo.id))
 }
 
@@ -51,7 +54,7 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req
   try {
-    const video = await Video.findById(id)
+    const video = await Video.findById(id).populate('creator')
     res.render('detailVideo', { pageTitle: video.title, video })
   } catch (error) {
     console.log(error)
@@ -64,7 +67,11 @@ export const getEditVideo = async (req, res) => {
   } = req
   try {
     const video = await Video.findById(id)
-    res.render('editVideo', { pageTitle: `Edit ${video.title}`, video })
+    if (video.creator !== req.user.id) {
+      throw Error()
+    } else {
+      res.render('editVideo', { pageTitle: `Edit ${video.title}`, video })
+    }
   } catch (error) {
     console.log(error)
     res.redirect(routes.home)
@@ -90,7 +97,11 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req
   try {
-    await Video.findByIdAndRemove({ _id: id })
+    if (video.creator !== req.user.id) {
+      throw Error()
+    } else {
+      await Video.findByIdAndRemove({ _id: id })
+    }
   } catch (error) {
     console.log(error)
   }
